Extend project details function to also include build system
OSS-Fuzz allows a variety of build systems. We can try to identify
which system a project uses by looking for keywords in build.sh.
To do this accurately for all projects, we'd need an exhaustive
list of allowed build systems and more refined logic.

As a starting point, I'm limiting myself to the build systems
mentioned by David and Adam in the article below. I imagine they
used those as example because they're common cases:

https://adalogics.com/blog/fuzzing-100-open-source-projects-with-oss-fuzz

One way to improve this functionality would be to read through the
build.sh documentation specific to each language (TODO) - e.g.:

https://google.github.io/oss-fuzz/getting-started/new-project-guide/python-lang/#buildsh