Ensure project cache uses original project names as keys
Users will know projects by their repo name and will naturally
expect to be able to find them by that exact name, dashes and all.

Internally, we can use sanitized versions of the names when needed,
most notably, when fetching data via GitHub's GraphQL API, since it
requires strict identifiers: no dashes, no periods, etc. But for
cache indexing, we can use the original names to facilitate things.

This was also the opportunity to introduce a dedicated module for
utility functions, which was bound to happen.