Add fallback logic in case GitHub's APIs can't be reached
This allows the prototype to continue to work when the external
APIs aren't reachable. In that case, it sources the data from a
local snapshot. It even works when completely offline.

The current fallback logic is just a temporary measure to ensure
the prototype can be reviewed no matter the contingencies. It
could be refactored into a proper solution, but that would require
more examination. (TODO)